Fixed user to objects mapping

+++ src/main/java/com/drawingreferencesaver/services/ReferenceService.java
@@ -1,10 +1,13 @@
 package com.drawingreferencesaver.services;
 
+import com.drawingreferencesaver.models.User;
 import com.drawingreferencesaver.repositories.ReferenceRepository;
 import com.drawingreferencesaver.models.Reference;
+import com.drawingreferencesaver.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.security.Principal;
 import java.util.List;
 import java.util.Optional;
 
@@ -12,6 +15,9 @@ import java.util.Optional;
 public class ReferenceService {
     @Autowired
     private ReferenceRepository referenceRepository;
+
+    @Autowired
+    private UserRepository userRepository;
 
     public List<Reference> list() {
         return referenceRepository.findAll();
@@ -21,19 +27,28 @@ public class ReferenceService {
         return this.referenceRepository.findById(id);
     }
 
-    public Reference addReference(Reference reference) {
+    public Reference addReference(Reference reference, Principal principal) {
+        User user = userRepository.findUsersByEmail(principal.getName());
+        reference.setUser(user);
         return this.referenceRepository.save(reference);
     }
 
-    public void deleteReference(long id) {
-        this.referenceRepository.deleteById(id);
+    public void deleteReference(long id, Principal principal) {
+        User user = userRepository.findUsersByEmail(principal.getName());
+        Reference reference = referenceRepository.getReferenceById(id);
+        if (reference.getUser().getId() == user.getId()) {
+            this.referenceRepository.deleteById(id);
+        }
     }
 
-    public Reference updateReference(Reference updatedReference, long id) {
+    public void updateReference(Reference updatedReference, long id, Principal principal) {
+        User user = userRepository.findUsersByEmail(principal.getName());
         Reference reference = referenceRepository.getReferenceById(id);
-        reference.setDescription(updatedReference.getDescription());
-        reference.setUrl(updatedReference.getUrl());
-        return referenceRepository.save(reference);
+        if (reference.getUser().getId() == user.getId()) {
+            reference.setDescription(updatedReference.getDescription());
+            reference.setUrl(updatedReference.getUrl());
+            referenceRepository.save(reference);
+        }
     }
 
 }
